fix: initially expand tree node for active file
When creating a new TreeNodeItem, expand it by default if a file
corresponding to that item (or one of its children, however deeply
nested) is active.

This is consistent with behavior when activating a file whose
TreeNodeItem already exists, and improves the experience when a story
file is discovered or added for any reason while its editor is already
open (e.g., the file becomes syntactically valid).
