+++ src/tree/TreeViewManager.ts
@@ -12,26 +12,7 @@ import { logWarn } from '../log/log';
 import type { StoryStore } from '../store/StoryStore';
 import { StoryTreeDataProvider } from './StoryTreeDataProvider';
 import type { TreeNode } from './TreeNode';
-
-const findEntryByUri = (
-  entries: Iterable<TreeNode>,
-  uri: Uri,
-): TreeNode | undefined => {
-  for (const entry of entries) {
-    if (entry.type === 'kind') {
-      if (entry.matchesUri(uri)) {
-        return entry;
-      }
-
-      const result = findEntryByUri(entry.getChildren(), uri);
-      if (result) {
-        return result;
-      }
-    }
-  }
-
-  return undefined;
-};
+import { findEntryByUri } from './findEntryByUri';
 
 const findEntryByStoryId = (
   entries: Iterable<TreeNode>,
